+++ src/Framework.php
@@ -6,9 +6,9 @@ namespace Framework;
 
 final class Framework
 {
-    public const string VERSION = '0.6.0';
+    public const string VERSION = '0.6.1';
 
-    public const array VERSION_TRIPLE = [0, 6, 0];
+    public const array VERSION_TRIPLE = [0, 6, 1];
 
     public const string VERSION_STABILITY = 'alpha';
 }